Добавил пару задач на Ресурсы и Логирование

+++ 09 Приведение примитивных типов расширение и сужение/task1007/src/task1007/Solution.java
@@ -23,7 +23,7 @@ Requirements:
 
 public class Solution {
     public static void main(String[] args) {
-        long l = (byte) 1234_564_890L;
+        long l =  1234_564_890L;
         int x = (byte) 0b1000_1100_1010;
         double m = (byte) 110_987_654_6299.123_34;
         float f = (byte) l++ + 10 + ++x - (float) m;
